Fix uninstall
