+++ src/Microsoft.DotNet.XHarness.Apple/Orchestration/UninstallOrchestrator.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Common.CLI;
@@ -59,34 +58,5 @@ namespace Microsoft.DotNet.XHarness.Apple
 
         protected override Task<ExitCode> InstallApp(AppBundleInformation appBundleInfo, IDevice device, TestTargetOs target, CancellationToken cancellationToken)
             => Task.FromResult(ExitCode.SUCCESS); // no-op - we only want to uninstall the app
-
-        private class FakeAppBundleInformationParser : IAppBundleInformationParser
-        {
-            private readonly string _bundleIdentifier;
-
-            public FakeAppBundleInformationParser(string bundleIdentifier)
-            {
-                _bundleIdentifier = bundleIdentifier ?? throw new ArgumentNullException(nameof(bundleIdentifier));
-            }
-
-            public Task<AppBundleInformation> ParseFromAppBundle(string appPackagePath, TestTarget target, ILog log, CancellationToken cancellationToken = default) =>
-                Task.FromResult(new AppBundleInformation(
-                    _bundleIdentifier,
-                    _bundleIdentifier,
-                    appPackagePath,
-                    appPackagePath,
-                    false));
-
-            public Task<AppBundleInformation> ParseFromProject(string projectFilePath, TestTarget target, string buildConfiguration)
-            {
-                var path = Path.GetDirectoryName(projectFilePath)!;
-                return Task.FromResult(new AppBundleInformation(
-                    _bundleIdentifier,
-                    _bundleIdentifier,
-                    path,
-                    path,
-                    false));
-            }
-        }
     }
 }
